requirements.txt
fastapi>=0.111.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
PyYAML>=6.0.1
uvicorn[standard]>=0.30.0
